move `durability` to its own module and make part of public API
