+++ src/derived/slot.rs
@@ -2,6 +2,7 @@ use crate::debug::TableEntry;
 use crate::dependency::DatabaseSlot;
 use crate::dependency::Dependency;
 use crate::derived::MemoizationPolicy;
+use crate::durability::Durability;
 use crate::lru::LruIndex;
 use crate::lru::LruNode;
 use crate::plumbing::CycleDetected;
@@ -9,7 +10,6 @@ use crate::plumbing::GetQueryTable;
 use crate::plumbing::HasQueryGroup;
 use crate::plumbing::QueryFunction;
 use crate::revision::Revision;
-use crate::runtime::Durability;
 use crate::runtime::FxIndexSet;
 use crate::runtime::Runtime;
 use crate::runtime::RuntimeId;
